Add validation and refactor models, repositories, and controllers for Category and Product
- Added spring-boot-starter-validation dependency in pom.xml
- Implemented Serializable interface in Category and Product models
- Updated ID types from int to Long in Product and Category models
- Refactored ProductRepository to use Long as ID type
- Enhanced CategoryController with CRUD operations and improved endpoint mappings
- Updated ProductController with refined endpoint mappings and CRUD operations
- Refactored CategoryServices and ProductServices to support new ID types and operations

+++ src/main/java/com/paterni/product/services/ProductServices.java
@@ -24,7 +24,7 @@ public class ProductServices {
         return productRepository.save(product);
     }
 
-    public Product getById(int id) {
+    public Product getById(long id) {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found"));
         return product;
@@ -34,12 +34,12 @@ public class ProductServices {
         return productRepository.findAll();
     }
 
-    public void deteleById(int id) {
+    public void deteleById(long id) {
         Product product = getById(id);
         productRepository.delete(product);
     }
 
-    public void update(int id , Product productUpdate) {
+    public void update(long id , Product productUpdate) {
     Product product = getById(id);
 
         checkCategory(productUpdate);
@@ -62,7 +62,7 @@ public class ProductServices {
        return categoryServices.getById(product.getCategory().getId());
     }
 
-    public Category getCategoryByProductID(int id){
+    public Category getCategoryByProductID(long id){
         return  getById(id).getCategory();
     }
 }
